+++ app/src/components/ProjectRegistration/ConfirmTransactionStep.svelte
