+++ app/src/components/ProjectRegistration/PickNameStep.svelte
