+++ app/src/components/ProjectRegistration/PickWalletStep.svelte
