+++ app/src/components/ProjectRegistration/TransactionSummaryStep.svelte
